feat : AttemptNumber 클래스 생성
시도 횟수 객체에서 검증까지 진행 한 후 컨트롤러에서 사용하는 방식으로 구현했습니다.
