--- src/main/java/racingcar/util/convertor/NumberConvertor.java
@@ -1,10 +1,0 @@
-package racingcar.util.convertor;
-
-import racingcar.util.validator.AttemptNumberValidator;
-
-public class NumberConvertor {
-	public static int convertNumber(String attemptNumber) {
-		AttemptNumberValidator.validateNumber(attemptNumber);
-		return Integer.parseInt(attemptNumber);
-	}
-}
